Separate TaskMap dnd and react-flow logic
Move ReactFlow to its own component to separate most drag-n-drop and
react-flow logic.

Furthermore, possible use of react-flow's useZoomPanHelper-hook is taken
into consideration in this decision.
(useZoomPanHelper provides utility functions for react-flow's viewport,
could be used to move canvas on drag)
useZoomPanHelper needs to be inside context of ReactFlowProvider
and being a hook, it needs to be used in a component.
