+++ frontend/src/components/TaskMap.module.scss
@@ -1,0 +1,30 @@
+@import './../shared.scss';
+
+$ZOOM: var(--zoom, 1);
+
+.taskMap {
+  @extend .layout-column;
+  width: 100%;
+}
+
+.taskName {
+  width: 210px;
+  max-height: 40px;
+  text-align: left;
+  overflow: hidden;
+  word-wrap: break-word;
+
+  -webkit-box-orient: vertical;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+
+  &.dragging {
+    width: calc(210px * #{$ZOOM});
+    max-height: calc(40px * #{$ZOOM});
+  }
+}
+
+.measureTaskName {
+  @extend .taskName;
+  visibility: hidden;
+}
