+++ frontend/src/components/TaskMapTask.module.scss
@@ -1,4 +1,4 @@
-@import '../pages/TaskMapPage.module.scss';
+@import './TaskMap.module.scss';
 
 .singleTask {
   @extend .layout-row;
