+++ frontend/src/components/TaskMapTaskGroup.module.scss
@@ -1,4 +1,4 @@
-@import '../pages/TaskMapPage.module.scss';
+@import './TaskMap.module.scss';
 
 .taskGroup {
   @extend .layout-column;
