+++ frontend/src/components/TaskRatingsText.module.scss
@@ -1,4 +1,4 @@
-@import '../pages/TaskMapPage.module.scss';
+@import './TaskMap.module.scss';
 
 .taskRatingRow {
   display: flex;
