+++ frontend/src/pages/TaskMapPage.module.scss
@@ -1,7 +1,5 @@
 @import '../shared.scss';
 
-$ZOOM: var(--zoom, 1);
-
 .taskmap {
   @extend .layout-column;
   height: 100%;
@@ -29,26 +27,4 @@ $ZOOM: var(--zoom, 1);
   left: 28px;
   bottom: 26px;
   width: 16px;
-}
-
-.taskName {
-  width: 210px;
-  max-height: 40px;
-  text-align: left;
-  overflow: hidden;
-  word-wrap: break-word;
-
-  -webkit-box-orient: vertical;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-
-  &.dragging {
-    width: calc(210px * #{$ZOOM});
-    max-height: calc(40px * #{$ZOOM});
-  }
-}
-
-.measureTaskName {
-  @extend .taskName;
-  visibility: hidden;
 }
